*** empty log message ***

+++ src/net/sf/zekr/engine/theme/ThemeTemplate.java
@@ -11,10 +11,8 @@ package net.sf.zekr.engine.theme;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.OutputStreamWriter;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Map.Entry;
+
+import org.apache.velocity.tools.generic.MathTool;
 
 import net.sf.zekr.common.runtime.Naming;
 import net.sf.zekr.engine.log.Logger;
@@ -25,21 +23,35 @@ import net.sf.zekr.engine.log.Logger;
  */
 public class ThemeTemplate extends BaseViewTemplate {
 
-	public String transform(ThemeData themeData) {
+	private ThemeData themeData;
+
+	public ThemeTemplate(ThemeData themeData) {
+		this.themeData = themeData;
+	}
+
+	/**
+	 * Transforms and persists the theme CSS file if doesn't exists in the cache (<tt>Naming.CACHE_DIR</tt>).
+	 * 
+	 * @return result CSS, or null if no transformation done
+	 */
+	public String transform() {
 		String retStr = null;
-		File srcFile = new File(resource.getString("theme.css", new String[] { themeData.id }));
-		File destFile = new File(Naming.CACHE_DIR + "/" + srcFile.getName());
+		File destFile = new File(Naming.CACHE_DIR + "/" + resource.getString("theme.css.fileName"));
 
-		themeData.process(config.getTranslation().getDefault().locale.getLanguage());
-		engine.putAll(themeData.processedProps);
+		// create destination CSS file if it doesn't exist
+		if (!destFile.exists() || destFile.length() == 0) {
+			File srcFile = new File(resource.getString("theme.css", new String[] { themeData.id }));
+			themeData.process(config.getTranslation().getDefault().locale.getLanguage());
+			engine.putAll(themeData.processedProps);
 
-		try {
-			OutputStreamWriter osw = new OutputStreamWriter(new FileOutputStream(destFile));
-			retStr = engine.getUpdated(srcFile.getPath());
-			osw.write(retStr);
-			osw.close();
-		} catch (Exception e) {
-			Logger.getLogger(this.getClass()).log(e);
+			try {
+				OutputStreamWriter osw = new OutputStreamWriter(new FileOutputStream(destFile));
+				retStr = engine.getUpdated(srcFile.getPath());
+				osw.write(retStr);
+				osw.close();
+			} catch (Exception e) {
+				Logger.getLogger(this.getClass()).log(e);
+			}
 		}
 		return retStr;
 	}
